Deleting MainController as it's not supposed to be used anymore (?). Deleting a bunch of stuff related to it such as Collection case class and functions in CalmService to retrieve collections from elasticsearch. Also moved some stuff out of common module that should sit in the api or in the calm adapter
