+++ api/src/main/scala/uk/ac/wellcome/platform/api/Server.scala
@@ -1,23 +1,16 @@
 package uk.ac.wellcome.platform.api
 
 import com.twitter.finagle.http.{Request, Response}
-
 import com.twitter.finatra.http.HttpServer
-import com.twitter.finatra.http.filters.{
-  CommonFilters,
-  LoggingMDCFilter,
-  TraceIdMDCFilter
-}
+import com.twitter.finatra.http.filters.{CommonFilters, LoggingMDCFilter, TraceIdMDCFilter}
 import com.twitter.finatra.http.routing.HttpRouter
 import com.twitter.finatra.json.modules.FinatraJacksonModule
 import com.twitter.finatra.json.utils.CamelCasePropertyNamingStrategy
-
 import uk.ac.wellcome.finatra.exceptions._
 import uk.ac.wellcome.finatra.modules._
-
 import uk.ac.wellcome.platform.api.controllers._
-
 import io.swagger.models.Swagger
+import uk.ac.wellcome.platform.api.finatra.exceptions.ElasticsearchExceptionMapper
 
 object ServerMain extends Server
 object ApiSwagger extends Swagger
@@ -51,7 +44,6 @@ class Server extends HttpServer {
       .filter[LoggingMDCFilter[Request, Response]]
       .filter[TraceIdMDCFilter[Request, Response]]
       .add[ManagementController]
-      .add[MainController]
       .add[ContextController]
       .add[SwaggerController]
       .add[WorksController]
