+++ api/src/main/scala/uk/ac/wellcome/platform/api/controllers/ManagementController.scala
@@ -2,15 +2,15 @@ package uk.ac.wellcome.platform.api.controllers
 
 import javax.inject.{Inject, Singleton}
 
+import com.sksamuel.elastic4s.ElasticClient
 import com.sksamuel.elastic4s.ElasticDsl._
 import com.twitter.finagle.http.Request
 import com.twitter.finatra.http.Controller
-import uk.ac.wellcome.finatra.services.ElasticsearchService
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 
 @Singleton
 class ManagementController @Inject()(
-  elasticsearchService: ElasticsearchService
+  elasticClient: ElasticClient
 ) extends Controller {
 
   get("/management/healthcheck") { request: Request =>
@@ -18,7 +18,7 @@ class ManagementController @Inject()(
   }
 
   get("/management/clusterhealth") { request: Request =>
-    elasticsearchService.client
+    elasticClient
       .execute { clusterHealth() }
       .map(health => response.ok.json(health.status))
   }
