+++ calm_adapter/src/main/scala/uk/ac/wellcome/calm_adapter/actors/DynamoRecordWriterActor.scala
@@ -8,11 +8,12 @@ import com.amazonaws.services.dynamodbv2.model._
 import com.google.inject.name.Named
 import com.gu.scanamo._
 import com.twitter.inject.Logging
-import uk.ac.wellcome.models.{ActorRegister, CalmTransformable}
+import uk.ac.wellcome.calm_adapter.models.ActorRegister
+import uk.ac.wellcome.models.CalmTransformable
 import uk.ac.wellcome.models.aws.DynamoConfig
 import uk.ac.wellcome.platform.calm_adapter.ServerMain
-import scala.concurrent.ExecutionContext.Implicits.global
 
+import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration.Duration
 
 /** Used to tell the OaiHarvestActor to slow down.
