+++ calm_adapter/src/main/scala/uk/ac/wellcome/calm_adapter/actors/OaiHarvestActor.scala
@@ -6,10 +6,10 @@ import akka.actor.{Actor, ActorSystem, PoisonPill}
 import akka.agent.Agent
 import com.google.inject.name.Named
 import com.twitter.inject.Logging
-import uk.ac.wellcome.models.ActorRegister
+import uk.ac.wellcome.calm_adapter.models.ActorRegister
 import uk.ac.wellcome.platform.calm_adapter.services._
-import scala.concurrent.ExecutionContext.Implicits.global
 
+import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration.Duration
 
 /** Configuration for the OAI harvest actor.
