--- common/src/main/scala/uk/ac/wellcome/models/Message.scala
@@ -1,7 +1,0 @@
-package uk.ac.wellcome.models
-
-trait Message {
-  val subject: Option[String]
-  val body: String
-  val topic: String
-}
