+++ common/src/main/scala/uk/ac/wellcome/models/aws/SQSMessage.scala
@@ -1,6 +1,5 @@
 package uk.ac.wellcome.models.aws
 
-import uk.ac.wellcome.models.Message
 import com.fasterxml.jackson.annotation._
 
 case class SQSMessage(
@@ -9,4 +8,4 @@ case class SQSMessage(
   @JsonProperty("TopicArn") topic: String,
   @JsonProperty("Type") messageType: String,
   @JsonProperty("Timestamp") timestamp: String
-) extends Message
+)
